+++ nfc_core/interfaces/innerkits/src/tags/ndef_message.cpp
@@ -97,7 +97,7 @@ std::shared_ptr<NdefRecord> NdefMessage::MakeUriRecord(const std::string& uriStr
         if (!uriString.compare(0, gUriPrefix[i].size(), gUriPrefix[i])) {
             payLoad += (i & 0xFF);
             uri = uriString.substr(gUriPrefix[i].size());
-            DebugLog("prefer index .%d", i);
+            DebugLog("prefer index .%d", (int)i);
             break;
         }
     }
